Create Hackathon Final Droney.blend
This is our final submission for the GIIS Hackathon 2K21 by Anchita Kharbanda and Jaidev Balaji.

##Inspirations
Our inspirations were start-ups like Zipline that have used drones to deliver essential medical supplies in LEDCs like Ghana and Rwanda, at low costs and high efficiency relative to traditional, road-based delivery mediums.

## What it does
Our project is an autonomous delivery vehicle in the form of a drone that can deliver items in a multitude of locations and climatic conditions, without human contact. This allows it to be implemented at continued efficiency levels, regardless of governmental guidelines and lockdowns.
## How we built it
We designed the propulsion system (the 4 rotors on our own) ensuring they were capable of generating enough lift to get the drone flying with a maximum operational payload. The rotors also give our drones VTOL (Vertical Take-Off & Landing) capabilities, extending their domain of operations. We gave special emphasis to the landing struts to be able to withstand hard landings as well as any operational stress. We aim to use a carbon-fibre body to reduce production costs and keep the UAV as light as possible.
## Challenges we ran into
The design for the landing struts took lots of precision and analysis of past use-cases considering the distribution of the weight of our drone as well as offering minimum drag while in flight. The 4 struts have angled padding on the ends to allow for landings in not-so-flat locations without risking much damage to the drone and/or the package.
## Accomplishments that we're proud of
We are proud of making a viable drone model on our own while optimising it to our use-case. We are also extremely satisfied with the usage of LiDAR (Light Detection And Ranging) and TCAS-like (Traffic Collision Avoidance System) technologies that are the best technologies for obstacle detection and collision avoidance in the market. Lastly, we are proud of being able to come up with a system that can increase the productivity of man and efficiency of deliveries.
## What we learned
We learned about the vulnerabilities of supply chains in the current world and hope that our project can fulfil the needs and wants of many others who might have had certain products rendered "inaccessible" due to such supply chain constraints as well as events like the COVID-19 pandemic.

## What's next for Droney
We would look forward to any further modifications of our model as well as the permits/legalities regarding such operations and re-shaping our idea for further use-cases!